Move TokensMoveDialog components from admin to dashboard
Move TokensMoveDialog components and styles from admin module
to dashboard module
