+++ src/modules/admin/components/Tokens/index.ts
@@ -1,6 +1,6 @@
+export { default as TokensMoveDialog } from '../../../dashboard/components/TokensMoveDialog';
 export { default } from './Tokens';
 export { default as ColonyTokenEditDialog } from './ColonyTokenEditDialog';
 export { default as TokenMintForm } from './TokenMintForm';
 export { default as TokenMintDialog } from './TokenMintDialog';
-export { default as TokensMoveDialog } from './TokensMoveDialog';
 export { default as TokenList } from './TokenList';
