+++ src/modules/dashboard/components/TokensMoveDialog/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TokensMoveDialog';
